requirements.txt
annotated-types==0.6.0
certifi==2024.2.2
charset-normalizer==3.3.2
google-auth==2.29.0
google-cloud-storage==2.16.0
google-cloud-bigquery==3.22.0
pendulum==3.0.0
idna==3.7
marvin==2.3.3
modal==0.62.153
openai==1.27.0
pydantic==2.7.1
pydantic-core==2.18.2
requests==2.31.0
typing-extensions==4.11.0
urllib3==2.2.1
uv==0.1.42
